refactor(types,db): slice types.ts + ingest.ts into domain modules (facade-preserving)
Behavior-preserving decomposition of the two worst oversized modules
(finding C7):
- Replace src/lib/types.ts (1562 lines) with src/lib/types/ (16 domain
  modules + a barrel index that export *'s all of them). All ~190
  '@/lib/types' importers unchanged — the barrel keeps the public
  surface identical.
- Extract pure, state-free helper clusters from db/ingest.ts into
  ingest/parseHelpers.ts + ingest/merge.ts; ingest.ts keeps all 10
  public exports as a facade. All 13 importers unchanged.

C4 (dual-backend): verified the safe shared post-processing already
lives in shared modules used by both backends; the full MINDER_USE_DB
dispatch unification and the deeper reconcile/write ingest split are
DEFERRED (module-state coupling would force circular imports and risk
behavior change — not safe to do unattended). Noted for a supervised
follow-up.

Gates: typecheck 0, parity + data/db/usage tests pass, build 0.
Finding from docs/reviews/2026-07-03-app-review.md.
