+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 ### Changed
+- **Internal: split oversized `types.ts` / `ingest.ts` into domain modules (facade-preserving).** `src/lib/types.ts` (~1560 lines) is sliced into cohesive domain modules under `src/lib/types/` (`project`, `session`, `board`, `ops`, `config`, `claudeConfig`, `lint`, `stats`, `template`, etc.) with `src/lib/types/index.ts` as a barrel that `export *`s every module, so the ~190 files importing `@/lib/types` are unchanged. Pure, module-state-free helpers in `src/lib/db/ingest.ts` (`extractFileOp`/`extractAgentName`/`extractSkillName`/`truncateText`/`readTailToLastNewline` → `ingest/parseHelpers.ts`; `safeExtractPrs`/`safeExtractTickets`/`mergePrLinks`/`mergeTicketLinks` → `ingest/merge.ts`) are extracted into sibling modules and re-imported; the public export surface of `@/lib/db/ingest` is unchanged. The cross-backend post-processing that could be shared between the SQLite and file-parse usage paths was already factored into shared helpers (`toLocalDateStr`, `groupByBinary`, `parseStoredArgs`); the remaining per-function divergences are intentional and left in place. No user-facing behavior change.
 - **RSC server-prefetch, Server Actions, and SSE live-updates are now enabled by default (were opt-in); bespoke pollers migrated to TanStack Query / the SSE bus.** The `rscHydration`, `serverActions`, and `liveEvents` feature flags — the shipped-but-dormant PR #237–#243 architecture — now default **on** (meta `defaultOn: true` and every call site reads `getFlag(..., key)` with no explicit `false`, so the Settings toggle and the runtime gate agree). Read-heavy pages now server-prefetch and hydrate, the two live writes route through Server Actions, and the client opens the single `/api/events` SSE stream by default; each flag can still be toggled **off** in **Settings**, which falls back to the intact client fetch-on-mount / POST-route / timer-poll paths. Alongside, seven hand-rolled `setInterval`+`fetch` pollers were migrated to `useQuery` with a matching `refetchInterval` and `refetchIntervalInBackground: false` so a backgrounded tab no longer polls: `useEfficiencyGrades`, `StatusDashboard` (3s), `DevServerControl` (2s, conditional), `SettingsPage` DB-health (15s), `BackgroundActivityBrowser` (15s); `useGitDirtyStatus` / `useGithubActivity` already ride the SSE bus by default and gained tab-visibility guards on their polling fallback. The two redundant 60s claude-status pollers (`ClaudeStatusBanner` + `ClaudeStatusListener`) are consolidated into a single `ClaudeStatusProvider` (one `useQuery` against a new combined `GET /api/claude-status/live` endpoint that returns snapshot + change events in one response) that feeds both the banner and the change-toast listener — one request instead of two.
 
 ### Added
